+++ model/item_category.gen.go
@@ -8,7 +8,7 @@ const TableNameItemCategory = "item_category"
 
 // ItemCategory mapped from table <item_category>
 type ItemCategory struct {
-	ID        int32  `gorm:"column:id;primaryKey;autoIncrement:true" json:"id"`
+	ID        int64  `gorm:"column:id;primaryKey;autoIncrement:true" json:"id"`
 	CreatedAt string `gorm:"column:created_at;not null" json:"created_at"`
 	UpdatedAt string `gorm:"column:updated_at;not null" json:"updated_at"`
 	IsDeleted bool   `gorm:"column:isDeleted;not null" json:"isDeleted"`
